update
practice_1_1_a_sample_c++_program.cpp
Savitch_2018_Problem Solving with C++
Chapter 1 Introduction to Computers and C++ Programming  

Using your text editor, enter (that is, type in) the C++ program shown in
Display 1.8. Be certain to type the first line exactly as shown in Display 1.8.
In particular, be sure that the first line begins at the left-hand end of the
line with no space before or after the # symbol. Compile and run the
program. If the compiler gives you an error message, correct the program
and recompile the program. Do this until the compiler gives no error
messages. Then run your program.