Update and rename 1_2_practice_program_favorite_number.cpp to practice_1_2_favorite_number.cpp
practice_1_2_favorite_number.cpp

Savitch_2018_Problem Solving with C++
Chapter 1 Introduction to Computers and C++ Programming  


Modify the C++ program you entered in Practice Program 1. Change the
program so that it asks a user for their favourite number, writes it to the
screen and then goes on to do the same things that the program in Display
1.8 does. Create an int variable to store the number and read a value
from the user using cin. Print out the text “Your favourite number is:” and
the number that the user has entered. Be certain to add the symbols \n to
your output statement. If the user entered 42, you should print out “Your
favourite number is 42”. Recompile the changed program and run it.