Update and rename 1_3_add_numbers.cpp to practice_1_3_add_numbers.cpp
/*
practice_1_3_add_numbers.cpp

Savitch_2018_Problem Solving with C++
Chapter 1 Introduction to Computers and C++ Programming  


Further modify the C++ program that you already modified in Practice
Program 2. 

Ask the user for two numbers.

Store the two numbers the user entered in two int variables. 

Print out the sum of these variables. 

For example, if the user entered 3 and 5, print out, “The sum of 3 and 5 is 8”.

*/