Create self_test_37_gallons_to_liters.cpp
/*
self_test_37_gallons_to_liters.cpp

Savitch_2018_Problem Soving with C++
2 C++ Basics
p 129

37. Write a complete C++ program that 

asks the user for a number of gallons
and then outputs the equivalent number of liters. 

There are 3.78533 liters in a gallon. 
Use a declared constant.

Since this is just an exercise, you need not have any comments in your program.
*/