Create self_test_23_if_else_exam_passed.cpp
/*
self_test_23_if_else_exam_passed.cpp

Savitch_2018_Problem Soving with C++
2 C++ Basics
Programming Projects
p115

Write an if-else statement that 
outputs the word Passed provided the value of the variable exam is greater than or equal to 60 and the value of the variable programs_done is greater than or equal to 10. 

Otherwise, the if-else statement outputs the word Failed. 

The variables exam and programs_done are both of type int.
*/