Create self_test_22_savings_expenses.cpp
/*

self_test_22_savings_expenses.cpp

Savitch_2018_Problem Soving with C++
2 C++ Basics
Programming Projects
p115

Suppose savings and expenses are variables of type double that have been given values. 

Write an if-else statement that outputs the word Solvent, decreases the value of savings by the value of expenses, and sets the value of expenses to 0, provided that savings is at least as large as expenses. 

If, however, savings is less than expenses, the if-else  statement simply outputs the word Bankrupt and does not change the value of any variables.
*/