Create practice_2_2_power_of_number.cpp
/*
practice_2_2_power_of_number.cpp

Savitch_2018_Problem Soving with C++
Chp 2 C++ Basics 
p 136

Powers of numbers can be calculated by multiplying the number by itself
for as many times as the value of the exponent. For example, 2 raised to
the power 4 can be calculated by multiplying 2 by itself 4 times to get 16.
Write a program that:

1. inputs a double as the base number and an int as the exponent;

2. multiplies the base number by itself using a loop that repeats for the
number of times in the int;

3. outputs the exponential value calculated.

Use an if statement for the special case where the output is 1 if the int
value is 0. 

//later
For a more challenging version, deal with the case where the
exponent is negative.

*/