Create practice_2_3_treadmill_conversion.cpp
/*
practice_2_3_treadmill_conversion.cpp
Book:
    Savitch_2018_Problem Soving with C++
    2 C++ Basics
    p 136

Date:   13 June 2022

Author:    Eton Williams

Required:
3. Many treadmills output the speed of the treadmill in miles per hour
(mph) on the console, but most runners think of speed in terms of a
pace. A common pace is the number of minutes and seconds per mile instead of mph.

Write a program that 
starts with a quantity in mph and converts the quantity into minutes and seconds per mile.
As an example, the proper output for an input of 6.5 mph should be 9 minutes and 13.8 seconds per mile. 

If you need to convert a double to an int, which will discard any value after the decimal point, then you may use
intValue = static_cast<int>(dblVal);