Create project_2_15_calculate_thermal_expansion.cpp
/*
project_2_15_calculate_thermal_expansion.cpp

Book:
    Savitch_2018_Problem Soving with C++
    2 C++ Basics
    p 142

Date:   20 June 2022

Author:    Eton Williams

Required:
15. It is important to consider the effect of thermal expansion when building a
structure that must withstand changes in temperature. For example, a metal
beam will expand in hot temperatures. The additional stress could cause
the structure to fail. Similarly, a material will contract in cold temperatures.
The linear change in length of a material if it is allowed to freely expand is
described by the following equation:
delta_L= alpha * L0* delta_T

Here, L0 is the initial length of the material in meters, delta_L is the displacement
in meters, delta_T is the change in temperature in Celsius, and alpha is a coefficient
for linear expansion.

Write a program that inputs a, LΔ, and TΔ, then calculates and outputs the
linear displacement. If the displacement is positive then output that “The
material will expand by” the displacement in meters. 

If the displacement is negative then output that “The material will contract by” the displacement in meters. 

You shouldn’t output the displacement as a negative number.

Here are some values for alpha for different materials.
Aluminum 2.31 × 10−5
Copper 1.70 × 10−5
Glass 8.50 × 10−6
Steel 1.20 × 10−5

*/