Create project_2_5_people_allowed_in_room.cpp
/*
project_2_5_people_allowed_in_room.cpp


     AUTHOR :  Eton C. Williams
       DATE :  21 June 2022
       BOOK :  Savitch_2018_Problem Soving with C++
               Chapter 2: C++ Basics
               p 136           

REQUIREMENTS:
5. Write a program that determines whether a meeting room is in violation of
fire law regulations regarding the maximum room capacity. The program
will read in the maximum room capacity and the number of people attending
the meeting. If the number of people is less than or equal to the
maximum room capacity, the program announces that it is legal to hold
the meeting and tells how many additional people may legally attend. If
the number of people exceeds the maximum room capacity, the program
announces that the meeting cannot be held as planned due to fire regulations
and tells how many people must be excluded in order to meet the
fire regulations. For a harder version, write your program so that it allows
the calculation to be repeated as often as the user wishes. If this is a class
exercise, ask your instructor whether you should do this harder version.


*/