Create  project_02_04_useful_loan_v2.cpp
/*
 project_02_04_useful_loan_v2.cpp

     AUTHOR :  Eton C. Williams
       DATE :  21 June 2022
       BOOK :  Savitch_2018_Problem Soving with C++
               Chapter 2: C++ Basics
               p 138           

REQUIREMENTS:
4. Negotiating a consumer loan is not always straightforward. One form of
loan is the discount installment loan, which works as follows. Suppose a loan has a 

face value of $1,000, 
interest rate is 15%,  
duration is 18 months. 

interest is computed by multiplying the face value of $1,000 by 0.15, to yield $150.

Total interest owed = interest multiplied by the loan period of 1.5 years to yield $225 

That amount is immediately deducted from the face value, leaving the consumer with only $775.

Repayment is made in equal monthly installments based on the face value. 

So the monthly loan payment will be $1,000 divided by 18, which is $55.56.

This method of calculation may not be too bad if the consumer needs $775 dollars, but the calculation is a bit more complicated if the consumer needs $1,000.

*/