Create project_2_5_net_pay_after_deductables.cpp
/*
project_2_5_net_pay_after_deductables.cpp

     AUTHOR :  Eton C. Williams
       DATE :  21 June 2022
       BOOK :  Savitch_2018_Problem Soving with C++
               Chapter 3: More Flow Of Control
               p 139           

REQUIREMENTS:
6. An employee is paid at a rate of $16.78 per hour for the first 40 hours
worked in a week. Any hours over that are paid at the overtime rate of oneand-
one-half times that. From the worker’s gross pay, 6% is withheld for
Social Security tax, 14% is withheld for federal income tax, 5% is withheld
for state income tax, and $10 per week is withheld for union dues. If the
worker has three or more dependents, then an additional $35 is withheld
to cover the extra cost of health insurance beyond what the employer pays.
Write a program that will read in the number of hours worked in a week
and the number of dependents as input and will then output the worker’s
gross pay, each withholding amount, and the net take-home pay for the
week. For a harder version, write your program so that it allows the calculation
to be repeated as often as the user wishes. 
*/