Create project_absolute_02_12_ideal_body_weight.cpp
/*
project_absolute_02_12_ideal_body_weight.cpp

     AUTHOR :  Eton C. Williams
       DATE :  25 June 2022
       BOOK :  Absolute C++
               Chapter 2: C++ Basics
               p 43              

REQUIREMENTS:
12. A simple rule to estimate your ideal body weight is to allow 110 pounds for the first 5
feet of height and 5 pounds for each additional inch.

Write a program with a variable for the 
height of a person in feet and another variable for the 
additional inches and input values for these variables from the keyboard. 

Assume the person is at least 5 feet tall. 
For example, a person that is 6 feet and 3 inches tall would be represented with
a variable that stores the number 6 and another variable that stores the number 3.
Based on these values calculate and output the ideal body weight.

*/