Create practice_03_02_revolving_credit_account.cpp
/*
practice_03_02_revolving_credit_account.cpp

     AUTHOR :  Eton C. Williams
       DATE :  27 June 2022
       BOOK :  Savitch_2018_Problem Soving with C++
               Chapter 3: More Flow of Control
               p 204           

REQUIREMENTS:
2. Write a program to compute the interest due, total amount due, and
the minimum payment for a revolving credit account. 

The program accepts the account balance as input, then adds on the interest to get the total amount due. 

The rate schedules are the following: The interest is
1.5 percent on the first $1,000 and 1 percent on any amount over that.

The minimum payment is the total amount due if that is $10 or less;
otherwise, it is $10 or 10 percent of the total amount owed, whichever is
larger. 

Your program should include a loop that lets the user repeat this
calculation until the user says she or he is done.

*/