Create practice_03_03_astrological_predictions_2.cpp
/*
practice_03_03_astrological_predictions_2.cpp

     AUTHOR :  Eton C. Williams
       DATE :  28 June 2022
       BOOK :  Savitch_2018_Problem Soving with C++
               Chapter 3: More Flow of Control
               p 204           

REQUIREMENTS:
3. Write an astrology program. The 

user types in a birthday, and the program responds with the sign and horoscope for that birthday. The month may be entered as a number from 1 to 12. Then enhance your program so that if the birthday is only one or two days away from an adjacent sign,
the program announces that the birthday is on a “cusp” and also outputs
the horoscope for that nearest adjacent sign.

This program will have a long multiway branch. Make up a horoscope for each sign. 

Your program should include a loop that lets the user repeat this calculation until the user says she or he is done.

The horoscope signs and dates are: