Create project_01_01_long_distance_calls.cpp
/*
project_01_01_long_distance_calls.cpp

     AUTHOR :  Eton C. Williams
       DATE :  29 June 2022
       BOOK :  Savitch_2018_Problem Soving with C++
               Chapter 3: More Flow of Control
               p 204           

REQUIREMENTS:
1. Write a program that computes the cost of a long-distance call. The cost of
the call is determined according to the following rate schedule:

a. Any call started between 8:00 am and 6:00 pm, Monday through Friday,
is billed at a rate of $0.40 per minute.

b. Any call starting before 8:00 am or after 6:00 pm, Monday through
Friday, is charged at a rate of $0.25 per minute.

c. Any call started on a Saturday or Sunday is charged at a rate of $0.15
per minute.

The input will consist of the day of the week, the time the call started, and
the length of the call in minutes. The output will be the cost of the call. The
time is to be input in 24-hour notation, so the time 1:30 pm is input as
13:30

The day of the week will be read as one of the following pairs of character
values, which are stored in two variables of type char:
Mo Tu We Th Fr Sa Su

Be sure to allow the user to use either uppercase or lowercase letters or a
combination of the two. 

The number of minutes will be input as a value of type int. 
(You can assume that the user rounds the input to a whole number of minutes.) 

Your program should include a loop that lets the
user repeat this calculation until the user says she or he is done.

*/