Create project_03_02_solve_quadratic_equation.cpp
/*
project_03_02_solve_quadratic_equation.cpp

     AUTHOR :  Eton C. Williams
       DATE :  01 July 2022
       BOOK :  Savitch_2018_Problem Soving with C++
               Chapter 3: More Flow of Control
               p 206           

REQUIREMENTS:
2. (This Project requires that you know some basic facts about complex
numbers, so it is only appropriate if you have studied complex numbers
in some mathematics class.)

Write a C++ program that solves a quadratic equation to find its roots.
The roots of a quadratic equation a*x*x + b*x + c = 0

(where a is not zero) are given by the formula
(−b ± sqrt(b*b − 4*a*c)) / 2*a

The value of the discriminant (b2 − 4ac) determines the nature of roots.

If the value of the discriminant is zero, then the equation has a single real
root. If the value of the discriminant is positive then the equation has two
real roots. If the value of the discriminant is negative, then the equation
has two complex roots.

The program takes values of a, b, and c as input and outputs the roots.

Be creative in how you output complex roots. 

Include a loop that allows the
user to repeat this calculation for new input values until the user says she
or he wants to end the program.

*/