Create project_03_03_roman_numerals.cpp
/*
project_03_03_roman_numerals.cpp

     AUTHOR :  Eton C. Williams
       DATE :  2 July 2022
       BOOK :  Savitch_2018_Problem Soving with C++
               Chapter 3: More Flow of Control
               p 207           

REQUIREMENTS:
3. Write a program that accepts a year written as a four-digit Arabic (ordinary)
numeral and outputs the year written in Roman numerals. 

Important Roman numerals are V for 5, X for 10, L for 50, C for 100, D for 500, and
M for 1,000. 

Recall that some numbers are formed by using a kind of
subtraction of one Roman “digit”; for example, IV is 4 produced as V minus
I, XL is 40, CM is 900, and so on. A few sample years: MCM is 1900, MCML
is 1950, MCMLX is 1960, MCMXL is 1940, MCMLXXXIX is 1989. 

Assume the year is between 1000 and 3000. Your program should include a loop
that lets the user repeat this calculation until the user says she or he is done.

*/