Create project_03_05_declining_balance.cpp
/*
project_03_05_declining_balance.cpp

     AUTHOR :  Eton C. Williams
       DATE :  29 June 2022
       BOOK :  Savitch_2018_Problem Soving with C++
               Chapter 3: More Flow of Control
               p 206           

REQUIREMENTS:
5. Interest on a loan is paid on a declining balance, and hence a loan with an
interest rate of, say, 14 percent can cost significantly less than 14 percent
of the balance. Write a program that takes a

 loan amount and interest rate as input and then 

outputs the monthly payments and balance of the loan until the loan is paid off. 

Assume that the monthly payments are onetwentieth of the original loan amount, and that 

any amount in excess of the interest is credited toward decreasing the balance due. 

Thus, on a loan of $20,000, the payments would be $1,000 a month. If the interest rate
is 10 percent, then each month the interest is one-twelfth of 10 percent
of the remaining balance. The first month (10 percent of $20,000)/12,
or $166.67, would be paid in interest, and the remaining $833.33 would
decrease the balance to $19,166.67. The following month the interest
would be (10 percent of $19,166.67)/12, and so forth. 

Also have the program output the total interest paid over the life of the loan.

Finally, determine what simple annualized percentage of the original loan
balance was paid in interest. 

For example, if $1,000 was paid in interest on
a $10,000 loan and it took 2 years to pay off, then the annualized interest
is $500, which is 5 percent of the $10,000 loan amount. Your program
should allow the user to repeat this calculation as often as desired.

*/